+++ src/main/kotlin/minesweeper/domain/position/Position.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.position
 
 data class Position(
     private val x: Coordinate,
