+++ src/test/kotlin/minesweeper/domain/position/CoordinateTest.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.position
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
